fix: fix miniFL test

+++ tests/forge/mock.t.sol
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.19;
 
-import {BaseTest} from "./base/BaseTest.sol";
+import { BaseTest } from "./base/BaseTest.sol";
 
 contract MockTest is BaseTest {
-    function test_ShouldWork() external {
-        assertEq(uint(1), uint(1));
-    }
+  function test_ShouldWork() external {
+    assertEq(uint256(1), uint256(1));
+  }
 }
